- Corrected grfio's 'next' structure to be int rather than short, as short has a 32k file limit - Moonlit petals now always knocks to the back of the character rather than pushing the character away from it's center. - Pressure no longer does full damage to plants. - Corrected 'timer already exists' when using @reloadbattleconf - Mob casted skills all have range 9 by default unless monster_ai 0x400 is specified. - Spider Web can't be dispelled now. - Magic Rod no longer shows an animation when used, only when it triggers. - Ganbantein now only removes the specific tiles it touches, not the whole ground spell. - Corrected pc_set registry functions to handle null strings. - Optimized the unequip loop when someone changed gender. - Made the mob damage log hold unsigned int for the damage value.
git-svn-id: http://eathena.googlecode.com/svn/branches/stable@10356 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ src/common/grfio.c
@@ -63,7 +63,7 @@ typedef struct {
 	int		srclen_aligned;		//
 	int		declen;				// original size
 	int		srcpos;
-	short	next;
+	int	next;
 	int	cycle;
 	char	type;
 	char	fn[128-4*5];		// file name
